Rename test files to example
