+++ src/data/store/middleware.ts
@@ -1,7 +1,7 @@
 import { compose, Middleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import { isDevelopment } from '';
+import { isDevelopment } from 'env';
 
 export const composeEnhancers = isDevelopment ? composeWithDevTools( {} ) : compose;
 export const sagaMiddleware = createSagaMiddleware();
